requirements.txt
pystray
Pillow
spotipy
colorist
colorsys
Pylette
pyserial